### 2019-08-30 #337 @wolski
presubmission : fgczgseaora
Submitting Author: Witold Wolski (@wolski)
Repository:   https://github.com/protViz/fgczgseaora

---

```
Package: fgczgseaora
Title: ORA, sigORA and GSEA functionalities for proteomic data
Version: 0.1.0
Authors@R: c(person("Lucas", "Kook",
    email = "lucasheinrich.kook@uzh.ch",
    role = c("aut"),
    person("Witold E.", "Wolski",
    email = "wew@fgcz.ethz.ch",
    role = c("cre"),
    comment = c(ORCID = "0000-0002-6468-120X")))
Description: Implements gene pair signature based over representation
    analysis and gene set enrichment tools. Allows background specification
    for sigORA using the sigora package  without further ID mapping being necessary.
Depends: R (>= 3.5), WebGestaltR (>= 0.4)
Imports: AnnotationDbi,
    GO.db,
    S4Vectors,
    UpSetR,
    DT,
    org.Hs.eg.db,
    quantable,
    reactome.db,
    sigora,
    tidyverse
Suggests: knitr, rmarkdown, bookdown, slam, BiocStyle, docopt
biocViews: MassSpectrometryData, ReproducibleResearch
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

A plethora of R packages exist on CRAN and Bioconductor to perform over-representation
analysis (ORA) and gene set enrichment analysis (GSEA). However, consistency in the
underlying nomenclature for specific analyses and user friendly implementation is
still lacking. `fgczgseaora` aims at unifying ID mapping and enrichment analysis
in a syntactically coherent and intuitive way, while ensuring reproducibility of
results. `fgczgseaora` primarily consists of wrapper functions around the
`r CRANpkg("sigora")` and `r CRANpkg("WebGestaltR")` packages from CRAN and
`r CRANpkg("rmarkdown")` based reports for visualisation and contextualisation
 of analysis results.

Biologist and Bioinformaticians working with proteomics data.

See two bullet points above.
