### 2019-09-02 #337 @wolski
Hello @melvidoni

This is a "presubmission inquiry".
I would expect to find an answer to question:
Not sure your software fits?

in your reply.

See also this link if you are not sure what I mean.
https://github.com/ropensci/software-review/issues/new/choose

Have a great day
Witek
