### 2019-09-02 #337 @melvidoni
Thank you @wolski for bringing your package to rOpenSci. If this is a presubmission, then kindly add that to the title of the issue, as it is expected. I will discuss this package with the other editors.
