### 2019-09-03 #337 @melvidoni
Hello @wolski. We have discussed your presubmission, but we found that your package is out of scope, as it doesn't fit completely on the categories we have for now. I do hope you can find another venue, as the package is interesting and useful. Thank you for submitting, and apologies for the original confusion due to the issue name.
