### 2020-07-21 #337 @wolski
Hello @melvidoni,

I did notice that since my submission a new category was introduced i.e.

Which I believe our package fgcz.gsea.ora fit's rather well.
Could you please have a second look at our application?
