### 2020-07-22 #337 @geanders
Thank you for your continued interest, @wolski . I have re-evaluated, and this package would still fall out of scope, even with the introduction of the new category of "scientific software wrappers". This category captures packages that wrap software that is available outside of the R environment, but not other R packages. There are some more details on this in the [ROpenSci book's section on Aims and Scope](https://devguide.ropensci.org/policies.html#aims-and-scope):

> "**scientific software wrappers:** Packages that wrap non-R utility programs used for scientific research. These programs must be specific to research fields, not general computing utilities. Wrappers must be non-trivial, in that there must be significant added value above simple system() call or bindings, whether in parsing inputs and outputs, data handling, etc. Improved installation process, or extension of compatibility to more platforms, may constitute added value if installation is complex. This does not include wrappers of other R packages or C/C++ libraries that can be included in R packages. We strongly encourage wrapping open-source and open-licensed utilities - exceptions will be evaluated case-by-case, considering whether open-source options exist."

It looks like the `fgczgseaora` package does wrap other software, but these are R packages, and so it does not fit into this new category, which focuses on providing access to functionality otherwise not available in R. I apologize that I do not have better news for you, but I wish you good luck in continuing work on developing and promoting your package.
